--- src/pages/blog/index.tsx
@@ -1,8 +1,0 @@
-import { Flex } from '@chakra-ui/react';
-import React from 'react';
-
-const Blog: React.FC = () => {
-    return <></>;
-}
-
-export default Blog;
